The project divided into modules
